+++ packages/clean_framework/lib/clean_framework.dart
@@ -11,5 +11,6 @@ export 'package:clean_framework/src/utilities/deserializer.dart';
 export 'package:clean_framework/src/utilities/either.dart';
 export 'package:clean_framework/src/utilities/network_logger.dart';
 export 'package:clean_framework/src/widgets/widgets.dart';
+export 'package:equatable/equatable.dart';
 export 'package:flutter_riverpod/flutter_riverpod.dart'
     show Consumer, Provider, ProviderContainer;