+++ packages/clean_framework/lib/src/utilities/clean_framework_observer.dart
@@ -4,6 +4,7 @@ import 'package:clean_framework/src/core/external_interface/request.dart';
 import 'package:clean_framework/src/core/external_interface/response.dart';
 import 'package:clean_framework/src/core/use_case/helpers/input.dart';
 import 'package:clean_framework/src/providers/external_interface.dart';
+import 'package:meta/meta.dart';
 
 /// The class to observe failures, route changes and other events.
 class CleanFrameworkObserver {
@@ -25,14 +26,19 @@ class CleanFrameworkObserver {
 
   /// Called when an [error] is thrown by [ExternalInterface]
   /// for the given [request].
+  @mustCallSuper
   void onExternalError(
     Object externalInterface,
     Request request,
     Object error,
+    StackTrace stackTrace,
   ) {
     log(
-      error.toString(),
-      name: '${externalInterface.runtimeType}[${request.runtimeType}]',
+      'Error occurred while requesting "${request.runtimeType}" '
+      'for "${externalInterface.runtimeType}"',
+      name: 'Clean Framework',
+      stackTrace: stackTrace,
+      error: error,
     );
   }