+++ packages/clean_framework_rest/lib/src/rest_service_options.dart
@@ -1,3 +1,7 @@
+import 'dart:async';
+
+import 'package:meta/meta.dart';
+
 class RestServiceOptions {
   const RestServiceOptions({
     this.baseUrl = '',
@@ -7,12 +11,13 @@ class RestServiceOptions {
   final String baseUrl;
   final Map<String, String> _headers;
 
-  Map<String, String> buildHeaders() {
+  @protected
+  FutureOr<Map<String, String>> buildHeaders() {
     return {
       'Content-Type': 'application/json; charset=UTF-8',
       ..._headers,
     };
   }
 
-  Map<String, String> get headers => buildHeaders();
+  FutureOr<Map<String, String>> get headers => buildHeaders();
 }